set permissions
The problem was if the file is already there and doesn't have the permissions. The old scenario didn't take this into account and didn't set the permission (which was my case, don't know why, maybe because the script incorrectly set them).
In this case we set them always, even if the file exists.
Removed from inside each file "if", to avoiding running them inside the loop.